requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0
python-decouple>=3.8
dotenv~=0.9.9
python-dotenv~=1.1.1
langgraph~=0.6.6
loguru~=0.7.3